+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Color random mode
+- Background-color random mode
+### Changed
+- `README.md` updated
+- Test system modified
 ## [2.5] - 2026-02-10
 ### Added
 - `--intensity` argument